feat: adding tabs component and refactoring the file into smaller components
